rename map_tests.py to map_correlations.py
